add distortion/base.py
