+++ warpfield/telescope/distortion/sip.py
@@ -1,9 +1,9 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
-''' define distortion function by SIP convention '''
+''' Define distortion function by the SIP convention '''
 
 from dataclasses import dataclass, field
-from scipy.optimize import least_squares
+from .base import BaseDistortion
 import numpy as np
 
 
@@ -27,7 +27,7 @@ class Sip:
     B: np.ndarray
 
     def __post_init__(self):
-        self.center = np.array((0, 0))
+        self.center = np.array((0, 0)).reshape((2,1))
         assert self.order >= 0, \
           f'The polynomical order should be non-negative.'
         assert self.A.shape == (self.order+1,self.order+1), \
@@ -35,8 +35,13 @@ class Sip:
         assert self.B.shape == (self.order+1,self.order+1), \
           f'The shape of B matris should be ({self.order+1}, {self.order+1}).'
 
+    def normalize(self, position: np.ndarray):
+        ''' Normalize position '''
+        position = position.copy()
+        return position - self.center
+
     def apply(self, position: np.ndarray):
-        ''' modify xy-coordinates with the SIP function
+        ''' Modify xy-coordinates with the SIP function
 
         Arguments:
           position (ndarray):
@@ -46,11 +51,8 @@ class Sip:
         Returns:
           An ndarray instance contains modified coordinates.
         '''
-        position = np.array(position).reshape((2, -1))
         N = position.shape[1]
-        cx, cy = self.center
-        x0, y0 = position[0], position[1]
-        x, y = x0 - cx, y0 - cy
+        x, y = self.normalize(position)
 
         dx = np.zeros_like(x)
         tmp = np.zeros((self.order + 1, N))
@@ -70,12 +72,12 @@ class Sip:
         for m in np.arange(self.order + 1):
             dy += tmp[m] * x**m
 
-        return np.vstack((x0 + dx, y0 + dy))
+        return position + np.stack((dx, dy))
 
 
 @dataclass
-class SipMod(Sip):
-    ''' modified SIP convention with the distortion center
+class AltSip(Sip):
+    ''' SIP convention with the displaed distortion center
 
     Attributes:
       order (int):
@@ -94,69 +96,11 @@ class SipMod(Sip):
     def __post_init__(self):
         assert self.center.size == 2, \
           'The center position should have two elements.'
-        self.center = self.center.flatten()
+        self.center = np.array(self.center).reshape((2,1))
 
 
-class __BaseSipDistortion:
-    def __call__(self, position: np.ndarray):
-        ''' distortion function with SIP convention
-
-        This function converts _correct_ coordinates into _distorted_ coordinates.
-        The distorted coordinates are obtained by an interative method.
-
-        Arguments:
-          position (ndarray):
-              A numpy.array with the shape of (2, Nsrc). The first element
-              contains the x-positions, while the second element contains
-              the y-positions.
-
-        Returns:
-          A numpy.ndarray of the input coordinates.
-        '''
-        position = np.array(position).reshape((2, -1))
-        p0, x0, d = position, position.copy(), np.infty
-        for n in range(100):
-            x1 = x0 + (p0 - self.apply(x0))
-            f, d, x0 = d, np.square(x1 - x0).mean(), x1
-            if d < 1e-24: break
-            if abs(1 - f / d) < 1e-3 and d < 1e-16: break
-            assert np.isfinite(d), \
-              'Floating value overflow detected.'
-        else:
-            raise RuntimeError(f'Iteration not converged ({d})')
-        return x0
-
-    def __solve__(self, position: np.ndarray):
-        ''' distortion function with SIP convention
-
-        This function converts _correct_ coordinates into _distorted_ coordinates.
-        The distorted coordinates are obtained by a least square minimization.
-        Note that this method fails if the number of positions is too large.
-
-        Arguments:
-          position (ndarray):
-              A numpy.array with the shape of (2, Nsrc). The first element
-              contains the x-positions, while the second element contains
-              the y-positions.
-
-        Returns:
-          A numpy.ndarray of the input coordinates.
-        '''
-        p0 = np.array(position).flatten()
-        func = lambda x: p0 - self.apply(x.reshape((2, -1))).flatten()
-        result = least_squares(func,
-                               p0,
-                               loss='linear',
-                               ftol=1e-15,
-                               xtol=1e-15,
-                               gtol=1e-15)
-        assert result.success is True, \
-          'failed to perform a SIP inverse conversion.'
-        return result.x.reshape((2, -1))
-
-
-class SipDistortion(Sip, __BaseSipDistortion):
-    ''' distortion function with the SIP (simple imaging polynomical) convention
+class SipDistortion(Sip, BaseDistortion):
+    ''' Distortion function with the SIP convention
 
     Attributes:
       order (int):
@@ -171,8 +115,8 @@ class SipDistortion(Sip, __BaseSipDistortion):
     pass
 
 
-class SipModDistortion(SipMod, __BaseSipDistortion):
-    ''' distortion function with the modified SIP convention
+class AltSipDistortion(AltSip, BaseDistortion):
+    ''' Distortion function with the displaced SIP convention
 
     Attributes:
       order (int):
